+++ backend/app/DataParsers/UsersParserProviderX.php
@@ -59,7 +59,7 @@ class UsersParserProviderX implements UsersParserInterface
     /**
      * @inheritDoc
      */
-    public function getBalance(array $data): int
+    public function getBalance(array $data): float
     {
         return $data['parentAmount'];
     }
